Move code in cli.go into other files according to their functionality.
