+++ cli/handlers.go
@@ -1,0 +1,129 @@
+package cli
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/raedahgroup/dcrcli/walletrpcclient"
+	qrcode "github.com/skip2/go-qrcode"
+)
+
+func balance(walletrpcclient *walletrpcclient.Client, commandArgs []string) (*response, error) {
+	balances, err := walletrpcclient.Balance()
+	if err != nil {
+		return nil, err
+	}
+
+	res := &response{
+		columns: []string{
+			"Account",
+			"Total",
+			"Spendable",
+			"Locked By Tickets",
+			"Voting Authority",
+			"Unconfirmed",
+		},
+		result: make([][]interface{}, len(balances)),
+	}
+	for i, v := range balances {
+		res.result[i] = []interface{}{
+			v.AccountName,
+			v.Total,
+			v.Spendable,
+			v.LockedByTickets,
+			v.VotingAuthority,
+			v.Unconfirmed,
+		}
+	}
+
+	return res, nil
+}
+
+func send(walletrpcclient *walletrpcclient.Client, commandArgs []string) (*response, error) {
+	sourceAccount, err := getSendSourceAccount(walletrpcclient)
+	if err != nil {
+		return nil, err
+	}
+
+	destinationAddress, err := getSendDestinationAddress(walletrpcclient)
+	if err != nil {
+		return nil, err
+	}
+
+	sendAmount, err := getSendAmount()
+	if err != nil {
+		return nil, err
+	}
+
+	passphrase, err := getWalletPassphrase()
+	if err != nil {
+		return nil, err
+	}
+
+	result, err := walletrpcclient.SendFromAccount(sendAmount, sourceAccount, destinationAddress, passphrase)
+	if err != nil {
+		return nil, err
+	}
+
+	res := &response{
+		columns: []string{
+			"Result",
+			"Hash",
+		},
+		result: [][]interface{}{
+			[]interface{}{
+				"The transaction was published successfully",
+				result.TransactionHash,
+			},
+		},
+	}
+
+	return res, nil
+}
+
+func receive(walletrpcclient *walletrpcclient.Client, commandArgs []string) (*response, error) {
+	var recieveAddress uint32
+
+	// if no address passed in
+	if len(commandArgs) == 0 {
+
+		// display menu options to select account
+		var err error
+		recieveAddress, err = getSendSourceAccount(walletrpcclient)
+		if err != nil {
+			return nil, err
+		}
+	} else {
+		// if an address was passed in eg. ./dcrcli receive 0 use that address
+		x, err := strconv.ParseUint(commandArgs[0], 10, 32)
+		if err != nil {
+			return nil, fmt.Errorf("Error parsing account number: %s", err.Error())
+		}
+
+		recieveAddress = uint32(x)
+	}
+
+	r, err := walletrpcclient.Receive(recieveAddress)
+	if err != nil {
+		return nil, err
+	}
+
+	qr, err := qrcode.New(r.Address, qrcode.Medium)
+	if err != nil {
+		return nil, fmt.Errorf("Error generating QR Code: %s", err.Error())
+	}
+
+	res := &response{
+		columns: []string{
+			"Address",
+			"QR Code",
+		},
+		result: [][]interface{}{
+			[]interface{}{
+				r.Address,
+				qr.ToString(true),
+			},
+		},
+	}
+	return res, nil
+}
